Move xla_metadata_call implementation to jax/_src/xla_metadata.py
It lived in jax/experimental/scheduling_groups.py for historical reasons. Its
public import is now jax.experimental.xla_metadata (still reachable from
scheduling_groups for compatibility).
