+++ jax/_src/BUILD
@@ -1362,11 +1362,15 @@ pytype_strict_library(
     deps = [
         ":ad",
         ":api",
+        ":api_util",
         ":batching",
         ":config",
         ":core",
+        ":flattree",
         ":mlir",
+        ":partial_eval",
         ":tree_util",
+        ":util",
         ":xla_metadata_lib",
         "//jax/_src/lib",
     ],
