+++ jax/experimental/BUILD
@@ -600,17 +600,7 @@ pytype_strict_library(
     type_checking = "pyrefly",
     visibility = ["//visibility:public"],
     deps = [
-        "//jax/_src:ad",
-        "//jax/_src:api",
-        "//jax/_src:api_util",
-        "//jax/_src:batching",
-        "//jax/_src:core",
-        "//jax/_src:flattree",
-        "//jax/_src:mlir",
-        "//jax/_src:partial_eval",
-        "//jax/_src:tree_util",
-        "//jax/_src:util",
-        "//jax/_src/lib",
+        "//jax/_src:xla_metadata",
     ],
 )
 
@@ -760,7 +750,6 @@ pytype_strict_library(
     type_checking = "pyrefly",
     visibility = ["//visibility:public"],
     deps = [
-        ":scheduling_groups",
         "//jax",
         "//jax/_src:xla_metadata",
     ],
