+++ jax/experimental/scheduling_groups.py
@@ -12,210 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from functools import partial
-
-from jax._src import core
-from jax._src import dispatch
-from jax._src import linear_util as lu
-from jax._src.api_util import debug_info
-from jax._src.util import (safe_map, safe_zip, weakref_lru_cache, unzip2,
-                           split_list, subs_list)
-from jax._src import flattree as ft
-from jax._src.tree_util import tree_flatten, tree_unflatten
-from jax._src.interpreters import ad, mlir, partial_eval as pe, batching
-from jax._src.lib.mlir.dialects import func as func_dialect
-from jax._src.lib.mlir import ir
-
-map, unsafe_map = safe_map, map
-zip, unsafe_zip = safe_zip, zip
+from jax._src.xla_metadata import xla_metadata_call
 
 def scheduling_group(name):
   return xla_metadata_call(scheduling_group=name)
-
-def xla_metadata_call(f=None, **meta):
-  if f is None:
-    return lambda g: _xla_metadata_call(g, **meta)
-  return _xla_metadata_call(f, **meta)
-
-# TODO(yashkatariya): Figure out a way to reuse code with compute_on2_p, fused_p
-def _xla_metadata_call(fun, **meta):
-  def wrapped(*args, **kwargs):
-    dbg = debug_info('xla_metadata_call', fun, args, kwargs)
-    args_ft = ft.flatten((args, kwargs))
-    in_avals = args_ft.map(core.shaped_abstractify)
-    jaxpr, out_avals = pe.trace_to_jaxpr(fun, in_avals, dbg)
-    if any(isinstance(c, core.Tracer) for c in jaxpr.consts):
-      jaxpr, consts = pe.separate_consts(jaxpr)
-    else:
-      consts = []
-    outs_flat = xla_metadata_call_p.bind(*consts, *args_ft.vals, jaxpr=jaxpr,
-                                         **meta)
-    return tree_unflatten(out_avals.tree, outs_flat)
-  return wrapped
-
-xla_metadata_call_p = core.Primitive('xla_metadata_call')
-xla_metadata_call_p.multiple_results = True
-dispatch.simple_impl(xla_metadata_call_p)
-
-
-def _xla_metadata_call_abstract_eval(*in_avals, jaxpr, **meta):
-  return jaxpr.out_avals
-xla_metadata_call_p.def_abstract_eval(_xla_metadata_call_abstract_eval)
-
-
-def attr_get(x):
-  if isinstance(x, str):
-    return ir.StringAttr.get(x)
-  else:
-    raise NotImplementedError(f'mlir attr handler for {type(x)=}')
-
-def _xla_metadata_call_lowering(ctx, *args, jaxpr, **meta):
-  const_args_and_avals = core.jaxpr_const_args(jaxpr)
-  const_args, const_avals = unzip2(const_args_and_avals)
-  in_avals = (*const_avals, *jaxpr.in_avals)
-  func_op, output_types, effects = mlir.lower_called_computation(
-      "xla_metadata_call", jaxpr, ctx.module_context, len(const_args), in_avals,
-      ctx.avals_out, ctx.tokens_in)
-
-  symbol_name = func_op.name.value
-  flat_output_types, treedef = mlir.ir_tree_registry.flatten(output_types)
-  tokens = [ctx.tokens_in.get(eff) for eff in effects]
-  hoisted_const_values, _ = mlir.ir_tree_registry.flatten([
-      mlir.ir_constants(c, const_lowering=ctx.const_lowering, aval=aval)
-      for c, aval in const_args_and_avals
-  ])
-  args = (*ctx.dim_var_values, *tokens, *hoisted_const_values, *args)
-  flat_args, _ = mlir.ir_tree_registry.flatten(args)
-  call = func_dialect.CallOp(
-      flat_output_types, ir.FlatSymbolRefAttr.get(symbol_name),
-      flat_args)
-  call.operation.attributes['mhlo.frontend_attributes'] = ir.DictAttr.get(
-      {k: attr_get(v) for k, v in meta.items()})
-  out_nodes = treedef.unflatten(call.results)
-  tokens, out_nodes = split_list(out_nodes, [len(effects)])
-  tokens_out = ctx.tokens_in.update_tokens(mlir.TokenSet(dict(zip(effects, tokens))))
-  ctx.set_tokens_out(tokens_out)
-  return out_nodes
-mlir.register_lowering(xla_metadata_call_p, _xla_metadata_call_lowering)
-
-
-def _xla_metadata_call_batcher(axis_data, vals_in, dims_in, *, jaxpr, **meta):
-  batched_jaxpr, dims_out = batching.batch_jaxpr2(jaxpr, axis_data, dims_in)
-  outs = xla_metadata_call_p.bind(*vals_in, jaxpr=batched_jaxpr, **meta)
-  return outs, dims_out
-batching.fancy_primitive_batchers[xla_metadata_call_p] = _xla_metadata_call_batcher
-
-
-def _xla_metadata_call_jvp(primals, tangents, *, jaxpr, **meta):
-  nzs = [not isinstance(t, ad.Zero) for t in tangents]
-  jaxpr_jvp, out_nzs = ad.jvp_jaxpr(jaxpr, nzs, False)
-  nz_tangents = [t for t in tangents if not isinstance(t, ad.Zero)]
-  outs = xla_metadata_call_p.bind(*primals, *nz_tangents, jaxpr=jaxpr_jvp, **meta)
-  primals_out, nz_tangents_out = outs[:len(out_nzs)], outs[len(out_nzs):]
-  nz_outs = iter(nz_tangents_out)
-  tangents_out = [next(nz_outs) if nz else ad.Zero(aval.to_tangent_aval())
-                  for aval, nz in zip(jaxpr.out_avals, out_nzs)]
-  assert next(nz_outs, None) is None
-  return primals_out, tangents_out
-ad.primitive_jvps[xla_metadata_call_p] = _xla_metadata_call_jvp
-
-
-def _xla_metadata_call_lin(is_vjp, nzs, *primals, jaxpr, **meta):
-  (primal_jaxpr, num_residuals_out, nzs_out, in_fwd_res,
-   tangent_jaxpr) = ad.linearize_jaxpr(jaxpr, nzs, is_vjp=is_vjp)
-
-  tangent_avals_out = [a.to_tangent_aval() for a in jaxpr.out_avals]
-
-  def _filter_zeros(is_nz_l, l):
-    return tuple(x for nz, x in zip(is_nz_l, l) if nz)
-
-  def tangent_fun(residuals, *tangents):
-    tangents_nz = _filter_zeros(nzs, tangents)
-    assert len(residuals) + len(tangents_nz) == len(tangent_jaxpr.invars), (
-        len(residuals), len(tangents_nz), len(tangent_jaxpr.invars))
-    nz_outs = xla_metadata_call_p.bind(*residuals, *tangents_nz,
-                                       jaxpr=tangent_jaxpr, **meta)
-    nz_outs_ = iter(nz_outs)
-    outs = [next(nz_outs_) if nz else ad.Zero(a)
-            for nz, a in zip(nzs_out, tangent_avals_out)]
-    assert next(nz_outs_, None) is None
-    return outs
-
-  ans = xla_metadata_call_p.bind(*primals, jaxpr=primal_jaxpr, **meta)
-  primal_ans, residuals_ans = split_list(ans, [len(ans) - num_residuals_out])
-  residuals_ans = subs_list(in_fwd_res, [*jaxpr.consts, *primals], residuals_ans)
-  return primal_ans, nzs_out, residuals_ans, tangent_fun
-ad.primitive_linearizations[xla_metadata_call_p] = _xla_metadata_call_lin
-
-
-@weakref_lru_cache
-def _transpose_jaxpr(jaxpr, in_tree, in_avals, specs):
-  out_tree = None
-  def transposed(*in_flat):
-    nonlocal out_tree
-    primals_ctrefs, cts_in = tree_unflatten(in_tree, in_flat)
-    args = ad.unproject_accums(specs, primals_ctrefs)
-    ad.backward_pass3(jaxpr, False, jaxpr.consts, args, cts_in)
-    cts_out = [x.freeze() if isinstance(x, ad.ValAccum) else None for x in args]
-    cts_out, out_tree = tree_flatten(cts_out)
-    return cts_out
-  dbg = jaxpr.debug_info.with_unknown_names()
-  trans_jaxpr, _, consts = pe.trace_to_jaxpr_dynamic(
-      lu.wrap_init(transposed, debug_info=dbg), in_avals)
-  return trans_jaxpr.with_consts(consts), out_tree
-
-
-def _xla_metadata_call_transpose(cts_in, *args, jaxpr, **meta):
-  primals_ctrefs, specs = ad.project_accums(args)
-  in_flat, in_tree = tree_flatten((primals_ctrefs, cts_in))
-  in_avals = [core.typeof(x) for x in in_flat]
-  trans_jaxpr, out_tree = _transpose_jaxpr(jaxpr, in_tree, (*in_avals,), specs)
-
-  cts_out = xla_metadata_call_p.bind(*in_flat, jaxpr=trans_jaxpr, **meta)
-
-  for x, ct in zip(args, tree_unflatten(out_tree, cts_out)):
-    if isinstance(x, ad.ValAccum):
-      x.accum(ct)
-
-
-ad.fancy_transposes[xla_metadata_call_p] = _xla_metadata_call_transpose
-
-
-def _xla_metadata_call_partial_eval_custom_params_updater(
-    unks_in,
-    inst_in,
-    kept_outs_known,
-    kept_outs_staged,
-    num_res_out,
-    num_res_in,
-    params_known,
-    params_staged,
-):
-  return params_known, params_staged
-
-
-pe.partial_eval_jaxpr_custom_rules[xla_metadata_call_p] = partial(
-    pe.closed_call_partial_eval_custom_rule,
-    'jaxpr',
-    _xla_metadata_call_partial_eval_custom_params_updater,
-)
-
-
-def dce_jaxpr_xla_metadata_rule(used_outputs: list[bool], eqn: pe.JaxprEqn
-                                ) -> tuple[list[bool], pe.JaxprEqn | None]:
-  if not any(used_outputs) and not pe.has_effects(eqn):
-    return [False] * len(eqn.invars), None
-  dced_jaxpr, used_inputs = pe._cached_closed_call_dce(
-      eqn.params['jaxpr'], tuple(used_outputs))
-  new_params = dict(eqn.params, jaxpr=dced_jaxpr)
-  if not any(used_inputs) and not any(used_outputs) and not dced_jaxpr.effects:
-    return used_inputs, None
-  else:
-    new_invars = [v for v, used in zip(eqn.invars, used_inputs) if used]
-    new_effs = core.eqn_effects(dced_jaxpr, new_invars)
-    new_eqn = pe.new_jaxpr_eqn(
-        new_invars,
-        [v for v, used in zip(eqn.outvars, used_outputs) if used],
-        eqn.primitive, new_params, new_effs, eqn.source_info, eqn.ctx)
-    return used_inputs, new_eqn
-pe.dce_rules[xla_metadata_call_p] = dce_jaxpr_xla_metadata_rule
